initial migration to v0.2 API - kin replay and vision demo scripts working, but unit tests not changed yet
