--- flygym/__init__.py
@@ -1,7 +1,0 @@
-from gymnasium.envs.registration import register
-
-
-register(
-    id="nmf-mujoco-base-v0",
-    entry_point="flygym.envs:NeuroMechFlyMuJoCo",
-)
