--- flygym/envs/nmf_isaacgym.py
@@ -1,7 +1,0 @@
-import numpy as np
-import gymnasium as gym
-from gymnasium import spaces
-
-
-class NeuroMechFlyIsaacGym(gym.Env):
-    raise NotImplementedError
