--- flygym/envs/nmf_pybullet.py
@@ -1,7 +1,0 @@
-import numpy as np
-import gymnasium as gym
-from gymnasium import spaces
-
-
-class NeuroMechFlyPyBullet(gym.Env):
-    raise NotImplementedError
