+++ flygym/mujoco/arena/__init__.py
@@ -1,0 +1,3 @@
+from .base import BaseArena, FlatTerrain
+from .complex_terrain import GappedTerrain, BlocksTerrain, MixedTerrain
+from .sensory_environment import OdorArena
