+++ flygym/mujoco/preprogrammed.py
@@ -1,0 +1,75 @@
+from flygym.common import get_data_path
+from typing import List
+from flygym.mujoco.state import KinematicPose
+
+
+all_leg_dofs = [
+    f"joint_{side}{pos}{dof}"
+    for side in "LR"
+    for pos in "FMH"
+    for dof in [
+        "Coxa",
+        "Coxa_roll",
+        "Coxa_yaw",
+        "Femur",
+        "Femur_roll",
+        "Tibia",
+        "Tarsus1",
+    ]
+]
+
+all_tarsi_links = [
+    f"{side}{pos}Tarsus{i}" for side in "LR" for pos in "FMH" for i in range(1, 6)
+]
+
+
+def get_preprogrammed_pose(pose: str):
+    data_path = get_data_path("flygym", "data") / "pose" / f"pose_{pose}.yaml"
+    if not data_path.is_file():
+        raise ValueError(f"Pose {pose} does not exist.")
+    return KinematicPose.from_yaml(data_path)
+
+
+def get_collision_geometries(config: str = "all") -> List[str]:
+    if config == "legs":
+        return [
+            f"{side}{pos}{dof}_collision"
+            for side in "LR"
+            for pos in "FMH"
+            for dof in [
+                "Coxa",
+                "Femur",
+                "Tibia",
+                "Tarsus1",
+                "Tarsus2",
+                "Tarsus3",
+                "Tarsus4",
+                "Tarsus5",
+            ]
+        ]
+    elif config == "legs-no-coxa":
+        return [
+            f"{side}{pos}{dof}_collision"
+            for side in "LR"
+            for pos in "FMH"
+            for dof in [
+                "Femur",
+                "Tibia",
+                "Tarsus1",
+                "Tarsus2",
+                "Tarsus3",
+                "Tarsus4",
+                "Tarsus5",
+            ]
+        ]
+    elif config == "tarsi":
+        return [
+            f"{side}{pos}{dof}_collision"
+            for side in "LR"
+            for pos in "FMH"
+            for dof in ["Tarsus1", "Tarsus2", "Tarsus3", "Tarsus4", "Tarsus5"]
+        ]
+    elif config == "none":
+        return []
+    else:
+        raise ValueError(f"Unknown collision geometry configuration: {config}")
