+++ flygym/mujoco/vision/__init__.py
@@ -1,0 +1,1 @@
+from .retina import Retina, visualize_visual_input
