--- flygym/state/__init__.py
@@ -1,2 +1,0 @@
-from .state import BaseState, KinematicPose
-from .preprogrammed import stretched_pose, zero_pose, walking_pose
